requirement.txt
pillow
scipy==1.1.0
tqdm
path.py